requirements.txt
Django==3.1.3
djangorestframework==3.12.2
psycopg2-binary==2.8.6
python-dotenv==0.15.0
requests==2.25.0
django-easy-scraper==1.0.1

